requirements.txt
Pillow
piexif
tkcalendar